Addressed comments and rearranged dependencies
