+++ helix-core/src/test/java/org/apache/helix/integration/multizk/TestMultiZkConnectionConfig.java
@@ -19,16 +19,17 @@ package org.apache.helix.integration.multizk;
  * under the License.
  */
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
-import org.apache.helix.HelixAdmin;
-import org.apache.helix.HelixManagerProperty;
-import org.apache.helix.TestHelper;
+import org.apache.helix.*;
+import org.apache.helix.cloud.constants.CloudProvider;
 import org.apache.helix.integration.manager.ClusterControllerManager;
 import org.apache.helix.integration.manager.MockParticipantManager;
 import org.apache.helix.integration.task.MockTask;
+import org.apache.helix.manager.zk.HelixManagerStateListener;
 import org.apache.helix.manager.zk.ZKHelixAdmin;
+import org.apache.helix.manager.zk.ZKHelixManager;
+import org.apache.helix.model.CloudConfig;
 import org.apache.helix.model.InstanceConfig;
 import org.apache.helix.participant.StateMachineEngine;
 import org.apache.helix.task.TaskFactory;
@@ -51,7 +52,11 @@ import org.testng.annotations.Test;
  * Tests were similar to TestMultiZkHelixJavaApis but without "MSDS_SERVER_ENDPOINT_KEY"
  * in system property
  */
-public class TestMultiZkConnectionConfig extends TestMultiZkHelixJavaApis {
+public class TestMultiZkConnectionConfig extends MultiZkTestBase {
+  protected ClusterSetup _clusterSetupZkAddr;
+  protected ClusterSetup _clusterSetupBuilder;
+  protected RealmAwareZkClient.RealmAwareZkConnectionConfig _invalidZkConnectionConfig;
+  protected RealmAwareZkClient.RealmAwareZkConnectionConfig _validZkConnectionConfig;
 
   @BeforeClass
   public void beforeClass() throws Exception {
@@ -59,7 +64,6 @@ public class TestMultiZkConnectionConfig extends TestMultiZkHelixJavaApis {
     // Routing data may be set by other tests using the same endpoint; reset() for good measure
     RoutingDataManager.getInstance().reset();
     // Create a FederatedZkClient for admin work
-
     try {
       _zkClient =
               new FederatedZkClient(new RealmAwareZkClient.RealmAwareZkConnectionConfig.Builder()
@@ -75,7 +79,27 @@ public class TestMultiZkConnectionConfig extends TestMultiZkHelixJavaApis {
     System.out.println("end start");
   }
 
-  @Override
+  /**
+   * Test cluster creation according to the pre-set routing mapping.
+   * Helix Java API tested is ClusterSetup in this method.
+   */
+  @Test
+  public void testCreateClusters() {
+    setupCluster();
+
+    createClusters(_clusterSetupZkAddr);
+    verifyClusterCreation(_clusterSetupZkAddr);
+
+    createClusters(_clusterSetupBuilder);
+    verifyClusterCreation(_clusterSetupBuilder);
+
+    // Create clusters again to continue with testing
+    createClusters(_clusterSetupBuilder);
+
+    _clusterSetupZkAddr.close();
+    _clusterSetupBuilder.close();
+  }
+
   public void setupCluster() {
     // Create two ClusterSetups using two different constructors
     // Note: ZK Address here could be anything because multiZk mode is on (it will be ignored)
@@ -85,6 +109,30 @@ public class TestMultiZkConnectionConfig extends TestMultiZkHelixJavaApis {
                             .setRoutingDataSourceEndpoint(_msdsEndpoint + "," + ZK_PREFIX + ZK_START_PORT)
                             .setRoutingDataSourceType(RoutingDataReaderType.HTTP_ZK_FALLBACK.name()).build())
             .build();
+  }
+
+  private void createClusters(ClusterSetup clusterSetup) {
+    // Create clusters
+    for (String clusterName : CLUSTER_LIST) {
+      clusterSetup.addCluster(clusterName, false);
+    }
+  }
+
+  private void verifyClusterCreation(ClusterSetup clusterSetup) {
+    // Verify that clusters have been created correctly according to routing mapping
+    _rawRoutingData.forEach((zkAddress, cluster) -> {
+      // Note: clusterNamePath already contains "/"
+      String clusterNamePath = cluster.iterator().next();
+
+      // Check with single-realm ZkClients
+      Assert.assertTrue(ZK_CLIENT_MAP.get(zkAddress).exists(clusterNamePath));
+      // Check with realm-aware ZkClient (federated)
+      Assert.assertTrue(_zkClient.exists(clusterNamePath));
+
+      // Remove clusters
+      clusterSetup
+              .deleteCluster(clusterNamePath.substring(1)); // Need to remove "/" at the beginning
+    });
   }
 
 
@@ -97,7 +145,6 @@ public class TestMultiZkConnectionConfig extends TestMultiZkHelixJavaApis {
   public void testCreateParticipants() throws Exception {
     // Create two ClusterSetups using two different constructors
     // Note: ZK Address here could be anything because multiZk mode is on (it will be ignored)
-    //HelixAdmin helixAdminZkAddr = new ZKHelixAdmin(ZK_SERVER_MAP.keySet().iterator().next());
     RealmAwareZkClient.RealmAwareZkConnectionConfig zkConnectionConfig =
             new RealmAwareZkClient.RealmAwareZkConnectionConfig.Builder()
                     .setRoutingDataSourceEndpoint(_msdsEndpoint + "," + ZK_PREFIX + ZK_START_PORT)
@@ -109,7 +156,6 @@ public class TestMultiZkConnectionConfig extends TestMultiZkHelixJavaApis {
 
     String participantNamePrefix = "Node_";
     int numParticipants = 5;
-    //createParticipantsAndVerify(helixAdminZkAddr, numParticipants, participantNamePrefix);
     createParticipantsAndVerify(helixAdminBuilder, numParticipants, participantNamePrefix);
 
     // Create mock controller and participants for next tests
@@ -154,8 +200,42 @@ public class TestMultiZkConnectionConfig extends TestMultiZkHelixJavaApis {
                       TestHelper.WAIT_DURATION));
     }
 
-    //helixAdminZkAddr.close();
     helixAdminBuilder.close();
+  }
+
+  protected void createParticipantsAndVerify(HelixAdmin admin, int numParticipants,
+                                             String participantNamePrefix) {
+    // Create participants in clusters
+    Set<String> participantNames = new HashSet<>();
+    CLUSTER_LIST.forEach(cluster -> {
+      for (int i = 0; i < numParticipants; i++) {
+        String participantName = participantNamePrefix + i;
+        participantNames.add(participantName);
+        InstanceConfig instanceConfig = new InstanceConfig(participantNamePrefix + i);
+        admin.addInstance(cluster, instanceConfig);
+      }
+    });
+
+    // Verify participants have been created properly
+    _rawRoutingData.forEach((zkAddress, cluster) -> {
+      // Note: clusterNamePath already contains "/"
+      String clusterNamePath = cluster.iterator().next();
+
+      // Check with single-realm ZkClients
+      List<String> instances =
+              ZK_CLIENT_MAP.get(zkAddress).getChildren(clusterNamePath + "/INSTANCES");
+      Assert.assertEquals(new HashSet<>(instances), participantNames);
+
+      // Check with realm-aware ZkClient (federated)
+      instances = _zkClient.getChildren(clusterNamePath + "/INSTANCES");
+      Assert.assertEquals(new HashSet<>(instances), participantNames);
+
+      // Remove Participants
+      participantNames.forEach(participant -> {
+        InstanceConfig instanceConfig = new InstanceConfig(participant);
+        admin.dropInstance(clusterNamePath.substring(1), instanceConfig);
+      });
+    });
   }
 
   /**
@@ -163,10 +243,50 @@ public class TestMultiZkConnectionConfig extends TestMultiZkHelixJavaApis {
    */
   @Test(dependsOnMethods = "testCreateParticipants")
   public void testZKHelixManager() throws Exception {
-    super.testZKHelixManager();
+    String clusterName = "CLUSTER_1";
+    String participantName = "HelixManager";
+    InstanceConfig instanceConfig = new InstanceConfig(participantName);
+    _zkHelixAdmin.addInstance(clusterName, instanceConfig);
+
+    createZkConnectionConfigs(clusterName);
+
+    HelixManagerProperty.Builder propertyBuilder = new HelixManagerProperty.Builder();
+    try {
+      HelixManager invalidManager = HelixManagerFactory
+              .getZKHelixManager(clusterName, participantName, InstanceType.PARTICIPANT, null,
+                      propertyBuilder.setRealmAWareZkConnectionConfig(_invalidZkConnectionConfig).build());
+      Assert.fail("Should see a HelixException here because the connection config doesn't have the "
+              + "sharding key set!");
+    } catch (HelixException e) {
+      // Expected
+    }
+
+    // Connect as a participant
+    HelixManager managerParticipant = HelixManagerFactory
+            .getZKHelixManager(clusterName, participantName, InstanceType.PARTICIPANT, null,
+                    propertyBuilder.setRealmAWareZkConnectionConfig(_validZkConnectionConfig).build());
+    managerParticipant.connect();
+
+    // Connect as an administrator
+    HelixManager managerAdministrator = HelixManagerFactory
+            .getZKHelixManager(clusterName, participantName, InstanceType.ADMINISTRATOR, null,
+                    propertyBuilder.setRealmAWareZkConnectionConfig(_validZkConnectionConfig).build());
+    managerAdministrator.connect();
+
+    // Perform assert checks to make sure the manager can read and register itself as a participant
+    InstanceConfig instanceConfigRead = managerAdministrator.getClusterManagmentTool()
+            .getInstanceConfig(clusterName, participantName);
+    Assert.assertNotNull(instanceConfigRead);
+    Assert.assertEquals(instanceConfig.getInstanceName(), participantName);
+    Assert.assertNotNull(managerAdministrator.getHelixDataAccessor().getProperty(
+            managerAdministrator.getHelixDataAccessor().keyBuilder().liveInstance(participantName)));
+
+    // Clean up
+    managerParticipant.disconnect();
+    managerAdministrator.disconnect();
+    _zkHelixAdmin.dropInstance(clusterName, instanceConfig);
   }
 
-  @Override
   protected void createZkConnectionConfigs(String clusterName) {
     RealmAwareZkClient.RealmAwareZkConnectionConfig.Builder connectionConfigBuilder =
             new RealmAwareZkClient.RealmAwareZkConnectionConfig.Builder();
@@ -185,6 +305,69 @@ public class TestMultiZkConnectionConfig extends TestMultiZkHelixJavaApis {
    */
   @Test(dependsOnMethods = "testZKHelixManager")
   public void testZKHelixManagerCloudConfig() throws Exception {
-    super.testZKHelixManagerCloudConfig();
+    String clusterName = "CLUSTER_1";
+    String participantName = "HelixManager";
+    InstanceConfig instanceConfig = new InstanceConfig(participantName);
+    _zkHelixAdmin.addInstance(clusterName, instanceConfig);
+
+    HelixManagerProperty.Builder propertyBuilder = new HelixManagerProperty.Builder();
+
+    // create a dummy cloud config and pass to ManagerFactory. It should be overwritten by
+    // a default config because there is no CloudConfig ZNode in ZK.
+    CloudConfig.Builder cloudConfigBuilder = new CloudConfig.Builder();
+    cloudConfigBuilder.setCloudEnabled(true);
+    // Set to Customized so CloudInfoSources and CloudInfoProcessorName will be read from cloud config
+    // instead of properties
+    cloudConfigBuilder.setCloudProvider(CloudProvider.CUSTOMIZED);
+    cloudConfigBuilder.setCloudID("TestID");
+    List<String> infoURL = new ArrayList<String>();
+    infoURL.add("TestURL");
+    cloudConfigBuilder.setCloudInfoSources(infoURL);
+    cloudConfigBuilder.setCloudInfoProcessorName("TestProcessor");
+
+    CloudConfig cloudConfig = cloudConfigBuilder.build();
+    HelixCloudProperty oldCloudProperty = new HelixCloudProperty(cloudConfig);
+    HelixManagerProperty helixManagerProperty =
+            propertyBuilder.setRealmAWareZkConnectionConfig(_validZkConnectionConfig)
+                    .setHelixCloudProperty(oldCloudProperty).build();
+    // Cloud property populated with fields defined in cloud config
+    oldCloudProperty.populateFieldsWithCloudConfig(cloudConfig);
+    // Add some property fields to cloud property that are not in cloud config
+    Properties properties = new Properties();
+    oldCloudProperty.setCustomizedCloudProperties(properties);
+
+    class TestZKHelixManager extends ZKHelixManager {
+      public TestZKHelixManager(String clusterName, String participantName,
+                                InstanceType instanceType, String zkAddress, HelixManagerStateListener stateListener,
+                                HelixManagerProperty helixManagerProperty) {
+        super(clusterName, participantName, instanceType, zkAddress, stateListener,
+                helixManagerProperty);
+      }
+
+      public HelixManagerProperty getHelixManagerProperty() {
+        return _helixManagerProperty;
+      }
+    }
+    // Connect as a participant
+    TestZKHelixManager managerParticipant =
+            new TestZKHelixManager(clusterName, participantName, InstanceType.PARTICIPANT, null, null,
+                    helixManagerProperty);
+    managerParticipant.connect();
+    HelixCloudProperty newCloudProperty =
+            managerParticipant.getHelixManagerProperty().getHelixCloudProperty();
+
+    // Test reading from zk cloud config overwrite property fields included in cloud config
+    Assert.assertFalse(newCloudProperty.getCloudEnabled());
+    Assert.assertNull(newCloudProperty.getCloudId());
+    Assert.assertNull(newCloudProperty.getCloudProvider());
+
+    // Test non-cloud config fields are not overwritten after reading cloud config from zk
+    Assert.assertEquals(newCloudProperty.getCustomizedCloudProperties(), properties);
+    Assert.assertEquals(newCloudProperty.getCloudInfoSources(), infoURL);
+    Assert.assertEquals(newCloudProperty.getCloudInfoProcessorName(), "TestProcessor");
+
+    // Clean up
+    managerParticipant.disconnect();
+    _zkHelixAdmin.dropInstance(clusterName, instanceConfig);
   }
 }
